+++ Busanify/Presentation/Views/Home/CustomViews.swift
@@ -138,3 +138,54 @@ class SearchTextField: UITextField {
         self.leftViewMode = .always
     }
 }
+
+class CustomCompass: UIButton {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configureButton()
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        configureButton()
+    }
+    
+    func configureButton() {        
+        self.translatesAutoresizingMaskIntoConstraints = false
+        self.widthAnchor.constraint(equalToConstant: 40).isActive = true
+        self.heightAnchor.constraint(equalToConstant: 40).isActive = true
+        self.backgroundColor = .white
+        
+        let stackView = UIStackView()
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        stackView.isUserInteractionEnabled = false
+        stackView.axis = .vertical
+        stackView.alignment = .center
+        stackView.spacing = 2
+        
+        if let triangleImage = UIImage(systemName: "triangle.fill") {
+            let triangleImageView = UIImageView(image: triangleImage)
+            triangleImageView.widthAnchor.constraint(equalToConstant: 15).isActive = true
+            triangleImageView.heightAnchor.constraint(equalToConstant: 15).isActive = true
+            triangleImageView.tintColor = .red
+            stackView.addArrangedSubview(triangleImageView)
+        }
+        
+        let label = UILabel()
+        label.text = "N"
+        label.textColor = .black
+        label.font = UIFont.systemFont(ofSize: 12)
+        stackView.addArrangedSubview(label)
+        
+        self.layer.cornerRadius = 20
+        self.layer.borderWidth = 0.5
+        self.layer.masksToBounds = true
+        
+        self.addSubview(stackView)
+        
+        NSLayoutConstraint.activate([
+            stackView.centerXAnchor.constraint(equalTo: self.centerXAnchor),
+            stackView.centerYAnchor.constraint(equalTo: self.centerYAnchor)
+        ])
+    }
+}
